name changes

+++ config_args.py
@@ -63,6 +63,8 @@ def get_args(parser):
 	parser.add_argument('-summarize_data', action='store_true')
 	parser.add_argument('-overwrite', action='store_true')
 	parser.add_argument('-int_preds', action='store_true')
+	parser.add_argument('-test_only', action='store_true')
+	parser.add_argument('-load_pretrained', action='store_true')
 	parser.add_argument('-int_pred_weight', type=float, default=0.2)
 	parser.add_argument('-matching_mlp', action='store_true')
 	parser.add_argument('-graph_conv', action='store_true')
@@ -80,8 +82,8 @@ def config_args(opt):
 	opt.multi_gpu = True
 	
 
-	if 'reuters' in opt.dataset or 'bibtext' in opt.dataset:
-		opt.results_dir = '/bigtemp/jjl5sw/deepENCODE/results/'
+	# if 'reuters' in opt.dataset or 'bibtext' in opt.dataset:
+	opt.results_dir = '/bigtemp/jjl5sw/deepENCODE/results/'
 
 	if opt.n_layers_dec is None:
 		opt.n_layers_dec = opt.n_layers_enc
@@ -253,7 +255,7 @@ def config_args(opt):
 		
 	print(opt.model_name)
 
-	if (not opt.viz) and (not opt.overwrite) and (not 'test' in opt.model_name) and (path.exists(opt.model_name)):
+	if (not opt.viz) and (not opt.overwrite) and (not 'test' in opt.model_name) and (path.exists(opt.model_name)) and (not opt.load_pretrained):
 	    overwrite_status = input('Already Exists. Overwrite?: ')
 	    if overwrite_status == 'rm':
 	    	os.system('rm -rf '+opt.model_name)
